Refactor, move remote classes to new package
