+++ app/src/main/java/com/tregula/graedukacyjna/composition/module/ConvertersModule.java
@@ -7,8 +7,8 @@ import com.tregula.graedukacyjna.domain.database.entity.Location;
 import com.tregula.graedukacyjna.domain.database.entity.converter.ContinentConverter;
 import com.tregula.graedukacyjna.domain.database.entity.converter.CountryConverter;
 import com.tregula.graedukacyjna.domain.database.entity.converter.LocationConverter;
-import com.tregula.graedukacyjna.domain.remote.ContinentJson;
-import com.tregula.graedukacyjna.domain.remote.CountryJson;
+import com.tregula.graedukacyjna.domain.remote.data.ContinentJson;
+import com.tregula.graedukacyjna.domain.remote.data.CountryJson;
 
 import javax.inject.Singleton;
