+++ app/src/main/java/com/tregula/graedukacyjna/domain/concurrent/action/FetchContinentsData.java
@@ -6,7 +6,7 @@ import com.google.android.gms.tasks.Task;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
 import com.squareup.moshi.Moshi;
-import com.tregula.graedukacyjna.domain.remote.ContinentsJson;
+import com.tregula.graedukacyjna.domain.remote.data.ContinentsJson;
 
 import java.io.IOException;
 import java.nio.charset.Charset;
