+++ app/src/main/java/com/tregula/graedukacyjna/domain/concurrent/action/PopulateDatabaseWithContinents.java
@@ -5,9 +5,9 @@ import com.tregula.graedukacyjna.domain.database.dao.ContinentDao;
 import com.tregula.graedukacyjna.domain.database.dao.CountryDao;
 import com.tregula.graedukacyjna.domain.database.entity.Continent;
 import com.tregula.graedukacyjna.domain.database.entity.Country;
-import com.tregula.graedukacyjna.domain.remote.ContinentJson;
-import com.tregula.graedukacyjna.domain.remote.ContinentsJson;
-import com.tregula.graedukacyjna.domain.remote.CountryJson;
+import com.tregula.graedukacyjna.domain.remote.data.ContinentJson;
+import com.tregula.graedukacyjna.domain.remote.data.ContinentsJson;
+import com.tregula.graedukacyjna.domain.remote.data.CountryJson;
 
 import java.util.ArrayList;
 import java.util.List;
