+++ app/src/main/java/com/tregula/graedukacyjna/domain/remote/data/CapitolJson.java
@@ -1,4 +1,4 @@
-package com.tregula.graedukacyjna.domain.remote;
+package com.tregula.graedukacyjna.domain.remote.data;
 
 public class CapitolJson {
     public String capitol_name;
