+++ app/src/main/java/com/tregula/graedukacyjna/domain/remote/data/ContinentJson.java
@@ -1,4 +1,4 @@
-package com.tregula.graedukacyjna.domain.remote;
+package com.tregula.graedukacyjna.domain.remote.data;
 
 import java.util.List;
